+++ CHANGELOG.md
@@ -5,3 +5,10 @@
 ## 0.0.2
 
 * Added copy-right header.
+
+## 0.0.3
+
+* Renamed:
+  * key_observer -> nd_keys_observer
+  * Name -> NDName
+* Added widget ```NDObserver```